+++ Projeto_Lista_Tarefas/index.html
@@ -21,7 +21,7 @@
   <header>
     <div class="header-section">
       <div class="logo-img">
-        <img src="/Projeto_Lista_Tarefas/Img/rocket.svg">
+        <img src="/Img/rocket.svg">
       </div>
       <div class="logo-text">
         <h1><span class="logo-color">to</span>do</h1>
@@ -37,7 +37,7 @@
 
         <input type="text" id="search" class="input-search" placeholder="Adicione uma nova tarefa" value="">
         <label for="search" class="label-search">
-          <button class="btn-search">Criar <img src="/Projeto_Lista_Tarefas/Img/plus.svg">
+          <button class="btn-search">Criar <img src="/Img/plus.svg">
           </button>
         </label>
 
@@ -65,7 +65,7 @@
 
         <div class="no-tasks">
           <div class="no-tasks-img">
-            <img src="/Projeto_Lista_Tarefas/Img/clipboard.svg">
+            <img src="/Img/clipboard.svg">
           </div>
 
           <div class="no-tasks-text">
@@ -88,7 +88,7 @@
     <h1>Feito por <a href="https://github.com/CiceroRMG">@cicerormg</a></h1>
   </footer>
 
-  <script src="/Projeto_Lista_Tarefas/scripts/script.js"></script>
+  <script src="/scripts/script.js" type="module"></script>
 </body>
 
 </html>
