+++ Projeto_Lista_Tarefas/scripts/create-task/Counter-Functions/addCreateTaskCount.js
@@ -1,0 +1,5 @@
+export function addCreateTaskCount() {
+  const count = document.querySelector(".task-new-conta");
+  const oldValue = Number(count.innerHTML);
+  count.innerHTML = `${oldValue + 1}`;
+}
