+++ Projeto_Lista_Tarefas/scripts/create-task/Counter-Functions/contador-Concluidas.js
@@ -1,0 +1,13 @@
+
+
+export function addContadorConcluidas() {
+  const contadorConcluidas = document.querySelector(".task-conclude-conta");
+  const oldValue = Number(contadorConcluidas.innerHTML);
+  contadorConcluidas.innerHTML = `${oldValue + 1}`;
+}
+
+export function subContadorConcluidas() {
+  const contadorConcluidas = document.querySelector(".task-conclude-conta");
+  const oldValue = Number(contadorConcluidas.innerHTML);
+  contadorConcluidas.innerHTML = `${oldValue - 1}`;
+}
