+++ Projeto_Lista_Tarefas/scripts/create-task/Counter-Functions/getCountTasksCreated.js
@@ -1,0 +1,4 @@
+export function getCountTasksCrated() {
+  const count = document.querySelector(".task-new-conta");
+  return Number(count.innerHTML);
+}
