+++ Projeto_Lista_Tarefas/scripts/create-task/Counter-Functions/subCreateTaskCount.js
@@ -1,0 +1,5 @@
+export function subCreateTaskCount() {
+  const count = document.querySelector(".task-new-conta");
+  const oldValue = Number(count.innerHTML);
+  count.innerHTML = `${oldValue - 1}`;
+}
