+++ Projeto_Lista_Tarefas/scripts/create-task/Trash-Button/createButtonTrash.js
@@ -1,0 +1,9 @@
+export function createButtonTrash() {
+  const button = document.createElement("button");
+  button.classList.add("btn-trash");
+  const imgTrash = document.createElement("img");
+  imgTrash.setAttribute("src", "/Img/trash.svg");
+  button.appendChild(imgTrash);
+
+  return button;
+}
