+++ Projeto_Lista_Tarefas/scripts/create-task/add-task-on-list.js
@@ -1,0 +1,4 @@
+export function addTaskOnList(task) {
+  const taskList = document.querySelector(".tasks-main-text");
+  taskList.appendChild(task);
+}
