+++ Projeto_Lista_Tarefas/scripts/create-task/create-task-node/createListItens.js
@@ -1,0 +1,5 @@
+export function createListItem() {
+  const createListItem = document.createElement("li");
+  createListItem.classList.add("tasks-main-form");
+  return createListItem;
+}
