+++ Projeto_Lista_Tarefas/scripts/create-task/no-taks-div/add-no-tasks.js
@@ -1,0 +1,4 @@
+export function addNoTaskDiv() {
+  const noTaskDiv = document.querySelector(".no-tasks-section");
+  noTaskDiv.classList.remove("hidden");
+}
