+++ Projeto_Lista_Tarefas/scripts/create-task/no-taks-div/hidden-no-tasks.js
@@ -1,0 +1,4 @@
+export function hiddenNoTaskDiv() {
+  const noTaskDiv = document.querySelector(".no-tasks-section");
+  noTaskDiv.classList.add("hidden");
+}
