+++ Projeto_Lista_Tarefas/scripts/create-task/select-search-input.js
@@ -1,0 +1,4 @@
+export function selectSearchInput() {
+  const input = document.querySelector(".input-search");
+  return input;
+}
