+++ Projeto_Lista_Tarefas/style.css
@@ -240,6 +240,11 @@ header span {
   height: 24px;
 }
 
+.btn-trash:hover {
+  background-color: #80808044;
+  border-radius: 999px;
+}
+
 .padding-buttom {
   margin-bottom: 100px;
 }
@@ -250,18 +255,17 @@ header span {
 }
 
 footer {
-  display: flex;
   margin: 20px auto;
-  align-items: center;
-  justify-content: center;
+  text-align: center;
+  margin-top: 20%;
 }
 
 footer h1, a {
   font-weight: 500;
-  font-size: 16px;
+  font-size: 14px;
   color: #D9D9D9;
 }
 
 footer a {
-  color: #4EA8DE;
+  color: #8284FA;
 }
